🚚 rename components file name
