--- src/components/layout/Layout.tsx
@@ -1,12 +1,0 @@
-import { Outlet } from "react-router-dom";
-import Header from "../header/Header";
-
-export default function Layout() {
-  return (
-    <div>
-      <Header />
-      THIS IS LAYOUT
-      <Outlet />
-    </div>
-  );
-}
